RentalManagement/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.93.5
drf-yasg==1.21.7
dynaconf==3.2.6
filelock==3.13.3
fontawesomefree==6.5.1
frozenlist==1.4.1
hvac==2.3.0
idna==3.8
inflection==0.5.1
knox==0.1.14
loguru==0.7.2
MarkupSafe==2.1.5
multidict==6.1.0
myenv==201902.7
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
phonenumbers==8.13.40
pillow==10.4.0
pipenv==2023.12.1
pipfile==0.0.2
platformdirs==4.2.0
propcache==0.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
toml==0.10.2
twilio==9.3.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
validators==0.33.0
virtualenv==20.25.1
Werkzeug==3.0.1
win32-setctime==1.1.0
yarl==1.15.4
